+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/HoneyBottleItemMixin.java
@@ -1,15 +1,15 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
 import insane96mcp.iguanatweaksreborn.module.Modules;
+import net.minecraft.world.item.HoneyBottleItem;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.PotionItem;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(PotionItem.class)
-public class PotionItemMixin {
+@Mixin(HoneyBottleItem.class)
+public class HoneyBottleItemMixin {
 	@Inject(at = @At("RETURN"), method = "getUseDuration", cancellable = true)
 	public void getUseDuration(ItemStack stack, CallbackInfoReturnable<Integer> callbackInfo) {
 		if (!Modules.hungerHealth.foodConsuming.fasterDrinkConsuming || !Modules.hungerHealth.foodConsuming.isEnabled())